Replaced Vue with Angular (WIP)

+++ server/src/webserver/routes.ts
@@ -11,6 +11,8 @@ import { UserPayload } from '../models/UserPayload';
 import { Request, Response } from 'express';
 import { ErrorTypes } from '../services/ErrorTypes';
 import { SoundMeta } from '../models/SoundMeta';
+import { ObjectID } from 'mongodb';
+import { GuildMember } from 'discord.js';
 
 export class Router {
   constructor(discordBot: DiscordBot, router: rs, fileHelper: FileHelper, databaseHelper: DatabaseHelper, private logger: Logger, authHelper: AuthHelper) {
@@ -34,11 +36,12 @@ export class Router {
         }
         catch (error) {
           this.logger.error(error, 'Login');
-          res.status(401).json({ message: 'Authentication failed' });
+          res.statusMessage = ErrorTypes.LOGIN_FAILED;
+          res.status(401).end();
         }
       });
 
-    router.route('/log/:serverId/:pageSize/:pageKey') // TO-DO: add parameter for serverId
+    router.route('/log/:serverId/:pageSize/:pageKey') // TO-DO: add parameter for serverId, pageSize and pageKey
       .get(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
         if (await discordBot.isUserAdminInServer(result.id, req.params.serverId)) {
@@ -61,12 +64,13 @@ export class Router {
         const result: UserPayload = this.getPayload(req);
         if (discordBot.isSuperAdmin(result.id)) {
           databaseHelper.udpateServerInfo(req.body.serverInfo)
-          res.status(200).json();
+          res.status(200).end();
         } else {
           const guild = await discordBot.isUserAdminInServer(result.id, req.body.serverInfo.id);
           if (guild) {
-            databaseHelper.udpateServerInfo(req.body.serverInfo)
-            res.status(200).json();
+            databaseHelper.udpateServerInfo(req.body.serverInfo);
+            res.statusMessage = 'Einstöllungen sen aufm neiestn Stond';
+            res.status(200).end();
           }
           else {
             this.notAdmin(res);
@@ -82,10 +86,12 @@ export class Router {
           const isAdmin: boolean = await discordBot.isUserAdminInServer(result.id, req.params.serverId) || discordBot.isSuperAdmin(result.id);
           try {
             await discordBot.playSoundCommand.stopPlaying(req.params.serverId, isAdmin);
-            res.status(200).json();
+            res.statusMessage = 'Da Bot holtat jetz de Goschn';
+            res.status(200).end();
           }
           catch (error) {
-            res.status(500).json(error);
+            res.statusMessage = error;
+            res.status(500).end();
           }
         }
         else {
@@ -97,22 +103,21 @@ export class Router {
       .get(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
         const servers = await discordBot.getUserServers(result.id);
-        res.status(200).json(databaseHelper.getSoundsMeta(servers));
+        res.status(200).json(databaseHelper.getSoundsMeta(servers.map(server => server.id)));
       });
 
     router.route('/sound/:soundId')
       .get(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
-        const meta = await databaseHelper.getSoundMeta(+req.params.soundId);
-        let status = false;
-        if (meta) {
-          status = await discordBot.isUserInServer(result.id, meta.serverId);
-        }
-
-        if (status) {
+        const soundId = new ObjectID(req.params.soundId);
+        const meta: SoundMeta | undefined = await databaseHelper.getSoundMeta(soundId);
+        
+        if (meta && await discordBot.isUserInServer(result.id, meta.serverId)) {
           res.status(200).download(meta.path, meta.fileName + path.extname(meta.path));
-        } else {
-          res.status(404).json();
+        }
+        else {
+          res.statusMessage = ErrorTypes.SOUND_NOT_FOUND;
+          res.status(404).end();
         }
       });
 
@@ -121,7 +126,16 @@ export class Router {
         const result: UserPayload = this.getPayload(req);
         const status = await discordBot.isUserInServer(result.id, req.params.serverId);
         if (status) {
-          res.status(200).json(databaseHelper.getSoundsMeta([{ id: req.params.serverId }]));
+          const soundMetas = await databaseHelper.getSoundsMeta([req.params.serverId], req.query.fromTime as string);
+          try{
+            await discordBot.mapUsernames(soundMetas, 'userId');
+            res.status(200).json(soundMetas);
+          }
+          catch (error) {
+            this.logger.error(error, 'Fetching user failed');
+            res.statusMessage = 'Konn de Sounds nit lodn. Frog Mr. Admin wos do vakehrt laft';
+            res.status(500).end();
+          }
         } else {
           this.notInServer(res, result.id, req.params.serverId);
         }
@@ -130,41 +144,40 @@ export class Router {
     router.route('/playSound')
       .post(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
-        const status = discordBot.isUserInServer(result.id, req.body.serverId);
+        const status = await discordBot.isUserInServer(result.id, req.body.serverId);
         if (status) {
           req.body.volume = Math.abs(req.body.volume); //negative values should not be possible
-          if (!discordBot.isSuperAdmin(result.id) && req.body.volume >= 1) {
+          if (!discordBot.isSuperAdmin(result.id) && req.body.volume > 1) {
             req.body.volume = 1;
           }
 
           const guild = await discordBot.getServer(req.body.serverId);
           const channelId = await discordBot.getChannelIdThroughUser(req.body.joinUser, req.body.serverId, req.body.channelId, guild, result.id);
-          try {
-            if (!channelId) throw ErrorTypes.USER_NOT_IN_VOICE_CHANNEL;
-
+          if (channelId) {
             if (req.body.soundId) {
               const meta = await databaseHelper.getSoundMeta(req.body.soundId);
               if (meta) {
                 databaseHelper.logPlaySound(result.id, guild.id, guild.name, meta);
-                let forcePlay = false;
-                if (req.body.forcePlay) {
-                  forcePlay = await discordBot.isUserAdminInServer(result.id, req.body.serverId)
-                }
+                const forcePlay = req.body.forcePlay ? await discordBot.isUserAdminInServer(result.id, req.body.serverId) : false;
+
                 await discordBot.playSoundCommand.requestSound(meta.path, req.body.serverId, channelId, req.body.volume, forcePlay);
-                res.status(200).json();
+                res.statusMessage = 'Sound sollt jetzt obgspült werdn';
+                res.status(200).end();
               }
               else {
-                res.status(404).json();
+                res.statusMessage = ErrorTypes.SOUND_NOT_FOUND;
+                res.status(404).end();
               }
             }
             else if (req.body.url) {
               await discordBot.playSoundCommand.playSound(req.body.serverId, channelId, undefined, req.body.volume, req.body.url);
-              res.status(200).json();
+              res.statusMessage = 'Sound sollt jetzt obgspült werdn';
+              res.status(200).end();
             }
           }
-          catch (error) {
-            this.logger.error(error, 'requestSound');
-            res.status(404).json(error);
+          else {
+            res.statusMessage = ErrorTypes.USER_NOT_IN_VOICE_CHANNEL;
+            res.status(404).end();
           }
         }
         else {
@@ -175,27 +188,31 @@ export class Router {
     router.route('/deleteSound/:id')
       .delete(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
-        const meta = await databaseHelper.getSoundMeta(+req.params.id);
-        if (meta?.userId === result.id || discordBot.isSuperAdmin(result.id)) {
+        const soundId = new ObjectID(req.params.id);
+        const meta = await databaseHelper.getSoundMeta(soundId);
+        if (meta && (meta.userId === result.id || discordBot.isSuperAdmin(result.id))) {
           const filePath = meta.path;
-          await databaseHelper.removeSoundMeta(+req.params.id, await discordBot.getServerName(meta.serverId));
+          await databaseHelper.removeSoundMeta(soundId);
           try {
             await fileHelper.deleteFile(filePath);
-            res.status(200).json();
+            res.statusMessage = 'Datei wurde eliminiert';
+            res.status(200).end();
           }
           catch (error) {
             this.logger.error(error, 'DeleteFile');
             try {
-              await databaseHelper.addSoundMeta(meta.id, meta.path, meta.fileName, meta.userId, meta.category, meta.serverId);
+              await databaseHelper.addSoundMeta(meta._id, meta.path, meta.fileName, meta.userId, meta.category, meta.serverId);
             }
             catch {
               this.logger.error(error, 'AddSoundMeta');
             }
-            res.status(500).json();
+            res.statusMessage = ErrorTypes.SOUND_COULD_NOT_BE_DELETED;
+            res.status(500).end();
           }
         }
         else {
-          res.status(403).json();
+          res.statusMessage = ErrorTypes.SOUND_NOT_FOUND;
+          res.status(404).end();
         }
       });
 
@@ -209,7 +226,7 @@ export class Router {
         }
         else {
           this.notInServer(res, result.id, req.params.serverId);
-        };
+        }
       });
 
     router.route('/soundCategories')
@@ -218,19 +235,20 @@ export class Router {
       });
 
     router.route('/uploadFile')
-      .put(upload.array('files'), async (req, res) => {
+      .post(upload.array('files'), async (req, res) => {
         const result: UserPayload = this.getPayload(req);
         const status = await discordBot.isUserInServer(result.id, req.body.serverId);
         if (status) {
-          let addedFiles = databaseHelper.addSoundsMeta(req.files, result.id, req.body.category, req.body.serverId);
-          res.status(200).json(addedFiles);
+          const newSoundMetas = await databaseHelper.addSoundsMeta(req.files, result.id, req.body.category, req.body.serverId);
+          res.statusMessage = 'Gratuliere! Du hosts gschofft a Datei hochzulodn :thumbsup:';
+          res.status(200).json(newSoundMetas);
         } else {
           try {
             await fileHelper.deleteFiles(req.files);
             this.notAdmin(res);
           }
           catch (error) {
-            this.logger.error(error, `DeleteFiles not possible.\nFiles: ${fileHelper.getFileNames(req.files).join('\n')}`);
+            this.logger.error(error, `DeleteFiles not possible.\nFiles: ${fileHelper.getFiles(req.files).join('\n')}`);
             this.notInServer(res, result.id, req.body.serverId);
           }
         }
@@ -239,10 +257,10 @@ export class Router {
     router.route('/setIntro')
       .post(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
-        const meta: SoundMeta = await databaseHelper.getSoundMeta(req.body.soundId);
-        const soundId = meta?.id ?? +req.body.soundId;
+        const meta: SoundMeta | undefined = await databaseHelper.getSoundMeta(req.body.soundId);
+        const soundId = meta?._id ?? req.body.soundId;
         let id: string;
-        let isAllowed: boolean = false;
+        let isAllowed = false;
 
         if (req.body.userId) {
           isAllowed = await discordBot.isUserAdminWhereAnotherUser(result.id, req.body.userId, req.body.serverId);
@@ -254,18 +272,20 @@ export class Router {
 
         if (isAllowed && (!meta || meta.serverId === req.body.serverId)) {
           databaseHelper.setIntro(id, soundId, req.body.serverId);
-          res.status(200).json();
+          res.statusMessage = 'Dei Untergong is gsetzt wordn';
+          res.status(200).end();
         }
         else {
-          res.status(500).json();
+          res.status(500).end();
         }
       });
 
     router.route('/user/:serverId')
       .get(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
-        const user = await discordBot.getUser(result.id);
-        res.status(200).json(databaseHelper.getUserInfo(user, req.params.serverId));
+        const intro = databaseHelper.getIntro(result.id, req.params.serverId);
+        await discordBot.mapUsernames([intro], 'id');
+        res.status(200).json(intro);
       });
 
     router.route('/hasAdminServers')
@@ -278,29 +298,28 @@ export class Router {
     router.route('/users/:serverId')
       .get(async (req, res) => {
         const result: UserPayload = this.getPayload(req);
+        let users: GuildMember[];
         if (discordBot.isSuperAdmin(result.id)) {
-          const users = await discordBot.getUsers(req.params.serverId);
-          res.status(200).json(databaseHelper.getUsersInfo(users, req.params.serverId));
+          users = await discordBot.getUsers(req.params.serverId);
+          
+        } else {
+          users = await discordBot.getUsersWhereIsAdmin(result.id);
         }
-        else {
-          const users = await discordBot.getUsersWhereIsAdmin(result.id);
-          if (users) {
-            res.status(200).json(databaseHelper.getUsersInfo(users, req.params.serverId));
-          }
-          else {
-            res.status(200).json([]);
-          }
-        }
+        const userInfo = await databaseHelper.getUsersInfo(users.map(user => user.id), req.params.serverId);
+        await discordBot.mapUsernames(userInfo, 'id');
+        res.status(200).json(userInfo);
       });
   }
 
   private notInServer(res: Response, userId: string, serverId: string) {
     this.logger.warn({ userId, serverId }, 'User not in server');
-    res.status(403).json({ message: 'Du host auf den Server kan Zugriff!!' });
+    res.statusMessage = 'Du host auf den Server kan Zugriff!!';
+    res.status(403).end();
   }
 
   private notAdmin(res: Response) {
-    res.status(403).json({ message: 'Nur der Hochadel hat Zugriff auf diese Funktion' });
+    res.statusMessage = 'Nur der Hochadel hat Zugriff auf diese Funktion';
+    res.status(403).end();
   }
 
   private getPayload(req: Request): UserPayload {
